overflow protect

+++ client/src/main/java/com/github/dtprj/dongting/remoting/NioWorker.java
@@ -110,7 +110,7 @@ public class NioWorker extends AbstractLifeCircle implements Runnable {
 
     @Override
     public void run() {
-        long cleanIntervalNanos = config.getCleanIntervalMills() * 1000 * 1000;
+        long cleanIntervalNanos = (long) config.getCleanIntervalMills() * 1000 * 1000;
         long lastCleanNano = System.nanoTime();
         int selectTimeoutMillis = config.getSelectTimeoutMillis();
         Selector selector = this.selector;
